"refactor: исправлены конструкторы User и Post"

+++ src/main/java/ru/petrojectgroup/conscience/service/PostService.java
@@ -2,13 +2,14 @@ package ru.petrojectgroup.conscience.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.petrojectgroup.conscience.exception.ValidationException;
 import ru.petrojectgroup.conscience.model.Post;
-import ru.petrojectgroup.conscience.storage.post.dao.PostStorage;
+import ru.petrojectgroup.conscience.storage.post.PostStorage;
 
 import java.util.Collection;
 
 @Service
-public class PostService implements PostStorage{
+public class PostService {
     private final PostStorage postStorage;
 
     @Autowired
@@ -20,16 +21,26 @@ public class PostService implements PostStorage{
         return postStorage.createPost(post);
     }
 
-    public void deletePost(long postId) {
-        postStorage.deletePost(postId);
+    public void deletePost(long postId, long userId) {
+        findPost(postId); //при неверном id поста выбросит исключение
+        if (checkPostAuthor(findPost(postId).getUserId(), userId)) {
+            postStorage.deletePost(postId);
+        } else {
+            throw new ValidationException("User is not author of the post");
+        }
     }
 
     public Collection<Post> findAll() {
         return postStorage.findAll();
     }
 
-    public Post updatePost(Post post) {
-        return postStorage.updatePost(post);
+    public Post updatePost(Post post, long userId) {
+        findPost(post.getPostId()); //при неверном id поста выбросит исключение
+        if (checkPostAuthor(post.getUserId(), userId)) {
+            return postStorage.updatePost(post);
+        } else {
+            throw new ValidationException("User is not author of the post");
+        }
     }
 
     public Post findPost(long postId) {
@@ -39,4 +50,9 @@ public class PostService implements PostStorage{
     public Collection<Post> findAllPostsOfUser(long userId) {
         return postStorage.findAllPostsOfUser(userId);
     }
+
+    private boolean checkPostAuthor(long userIdFromPost, long userId) {
+        return userIdFromPost == userId;
+    }
+    //TODO + проверка на юзеера
 }
